+++ fair_research_login/__init__.py
@@ -1,3 +1,4 @@
+import logging
 from fair_research_login.client import NativeClient
 from fair_research_login.token_storage import (ConfigParserTokenStorage,
                                                MultiClientTokenStorage,
@@ -6,7 +7,8 @@ from fair_research_login.token_storage import (ConfigParserTokenStorage,
 from fair_research_login.code_handler import (InputCodeHandler, CodeHandler)
 from fair_research_login.local_server import LocalServerCodeHandler
 from fair_research_login.exc import (LoginException, LoadError, ScopesMismatch,
-                                     TokensExpired)
+                                     TokensExpired, LocalServerError,
+                                     AuthFailure)
 
 __all__ = [
     'NativeClient',
@@ -17,5 +19,8 @@ __all__ = [
     'CodeHandler', 'InputCodeHandler', 'LocalServerCodeHandler',
 
     'LoginException', 'LoadError', 'ScopesMismatch', 'TokensExpired',
-    'LocalServerError',
+    'LocalServerError', 'AuthFailure',
 ]
+
+# https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library  # noqa
+logging.getLogger("fair_research_login").addHandler(logging.NullHandler())