+++ tests/unit/test_code_handler.py
@@ -1,10 +1,4 @@
 import pytest
-import sys
-
-try:
-    from unittest.mock import Mock
-except ImportError:
-    from mock import Mock
 
 from fair_research_login.code_handler import InputCodeHandler, CodeHandler
 
@@ -14,37 +8,34 @@ def test_code_handler_extendable_methods():
         CodeHandler().get_code()
 
 
-def test_code_handler_exits_on_interrupt(monkeypatch):
-
-    monkeypatch.setattr(sys, 'exit', Mock())
-
-    class MyCodeHandler(CodeHandler):
-        def get_code(self):
-            raise KeyboardInterrupt()
-
-    MyCodeHandler().authenticate('foo', no_browser=True)
-    assert sys.exit.called
-
-
 def test_code_handler_authenticate_with_webbrowser(mock_webbrowser,
                                                    mock_input,
                                                    mock_is_remote_session):
-    InputCodeHandler().authenticate('http://foo.edu', no_browser=False)
+    CodeHandler.set_browser_enabled(True)
+    InputCodeHandler().authenticate('http://foo.edu')
     assert mock_input.called
     assert mock_webbrowser.called
 
 
+def test_code_handler_set_browser_enabled_is_boolean():
+    with pytest.raises(ValueError):
+        CodeHandler.set_browser_enabled('invalid value')
+
+
 def test_code_handler_authenticate_without_webbrowser(mock_webbrowser,
                                                       mock_input):
-    InputCodeHandler().authenticate('http://foo.edu', no_browser=True)
+    CodeHandler.set_browser_enabled(False)
+    InputCodeHandler().authenticate('http://foo.edu')
     assert mock_input.called
     assert not mock_webbrowser.called
+    CodeHandler.set_browser_enabled(True)
 
 
 def test_code_handler_authenticate_with_ssh_session(mock_webbrowser,
                                                     mock_input,
                                                     monkeypatch):
+    CodeHandler.set_browser_enabled(True)
     monkeypatch.setenv('SSH_TTY', 'SSH_CONNECTION')
-    InputCodeHandler().authenticate('http://foo.edu', no_browser=False)
+    InputCodeHandler().authenticate('http://foo.edu')
     assert mock_input.called
     assert not mock_webbrowser.called